+++ CHANGELOG.md
@@ -1,0 +1,8 @@
+## 0.2.0
+
+- Load history positions of way after click 'Show More' button (to address [comment](https://github.com/osmlab/osm-deep-history/issues/14#issuecomment-335953135))
+- Change development server to `budo`
+
+## 0.1.0
+
+- Rewrite from osmlab/osm-deep-history with several enhancement
