+++ README.md
@@ -1,12 +1,12 @@
-# OSM deep history
+# OSM Visual History
 
 A better way to view the history of objects in [OpenStreetMap](http://www.openstreetmap.org/).
 
 Example history views:
-* [Node](https://aleung.github.io/osm-deep-history/#/node/4857559003)
-* [Way](https://aleung.github.io/osm-deep-history/#/way/333067739) (The map needs a few seconds to load)
+* [Node](https://aleung.github.io/osm-visual-history/#/node/4857559003)
+* [Way](https://aleung.github.io/osm-visual-history/#/way/333067739) (The map needs a few seconds to load)
 
-It's an ehancement base on [osmlab/osm-deep-history](https://github.com/osmlab/osm-deep-history) with below update:
+It's an enhancement base on [osmlab/osm-deep-history](https://github.com/osmlab/osm-deep-history) with below update:
 
 Common:
 * Improve map UI and fix issues
@@ -24,7 +24,7 @@ Way:
 
 ## Development
 
-The source code has been rewritten mostly in ES6, using Promise and functional style. 
+The source code has been rewritten mostly in ES6, using Promise and functional style.
 
 To rebuild and run locally:
