+++ index.html
@@ -10,7 +10,7 @@
 <body>
     <div class='container'>
         <div class='col12 pad1'>
-            <div class='col2'>
+            <div class='col1'>
                 <select id='type'>
                     <option value='node'>Node</option>
                     <option value='way'>Way</option>
@@ -23,7 +23,10 @@
             <div class='col1'>
                 <button id='go'>Get History</button>
             </div>
-            <div id='loading' class='col1'>
+            <div id='status' class='col2'>
+            </div>
+            <div class='col1'>
+                <button id='more' class='hide'>Show More on Map</button>
             </div>
         </div>
     </div>
