+++ package.json
@@ -1,17 +1,20 @@
 {
   "name": "osm-deep-history",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "description": "Display an OSM object's history",
   "main": "index.js",
   "scripts": {
-    "build": "browserify -g uglifyify js/site.js > js/bundle.js",
-    "start": "beefy index.js:bundle.js 8080 --browserify=./node_modules/browserify/bin/cmd.js -- -d"
+    "build": "browserify -g uglifyify index.js > js/bundle.js",
+    "start": "budo index.js:js/bundle.js --live --open"
   },
   "repository": "git@github.com:iandees/osm-deep-history.git",
   "keywords": [
     "openstreetmap"
   ],
   "author": "Ian Dees <ian.dees@gmail.com>",
+  "contributors": [
+    "Leo Liang <leoliang@gmail.com"
+  ],
   "license": "BSD",
   "dependencies": {
     "d3": "~4.10.2",
@@ -20,8 +23,8 @@
     "reqwest": "~2.0.5"
   },
   "devDependencies": {
-    "beefy": "2.1.8",
     "browserify": "13.0.1",
+    "budo": "^10.0.4",
     "uglifyify": "^4.0.4"
   }
 }
